requirements.txt
# IA — LLM
openai

# Web API
fastapi
uvicorn[standard]

# Audio (local CLI)
google-cloud-texttospeech
pygame
SpeechRecognition
pyaudio
requests
pyttsx3

# Auth / sessões persistentes
bcrypt
python-jose[cryptography]

# Utilitários
python-dotenv

# Testes
pytest
pytest-cov
